chore: Tidy up code and create renderInfoBox
