+++ svelte/src/scripts/draw/nodes-render.ts
@@ -1,6 +1,7 @@
 import * as d3 from 'd3';
 import {toHTMLToken} from '$helper';
 import type {DrawSettingsInterface, GraphDataNode} from '$types';
+import { renderInfoBox } from './helper/info-box';
 
 /**
  * Adds lift and collapse buttons to all groups with id .nodes (generated by renderNodes)
@@ -134,15 +135,8 @@ export function renderNodes(
 				d3.select(`#line-${toHTMLToken(link.id)}`).style('stroke-width', '4');
 			});
 
-			// Set the data for the info box
-			const infoBox = document.getElementById('info-box')!;
-			infoBox.style.display = 'block';
-			infoBox.style.background = drawSettings.nodeColors[level] ?? drawSettings.nodeDefaultColor;
-			infoBox.style.fillOpacity = '0.1';
-			document.getElementById('info-box-name')!.innerText = data.simpleName;
-			const subHeading = data.rs === '' ? data.kind : data.kind + ' - ' + data.rs;
-			document.getElementById('info-box-heading2')!.innerText = subHeading;
-			document.getElementById('info-box-desc')!.innerText = data.description;
+			// render info box
+			renderInfoBox(data, drawSettings.nodeColors[level] ?? drawSettings.nodeDefaultColor);
 		})
 		.on('mouseout', function (event, data) {
 			d3.select(this).attr('fill-opacity', '0.1');
